fix more debugger tests
