+++ packages/devtools_app/test/debugger/debugger_codeview_test.dart
@@ -104,7 +104,7 @@ void main() {
         findsOneWidget,
       );
       await expectLater(
-        find.byKey(DebuggerScreenBody.codeViewKey),
+        find.byType(CodeView),
         matchesDevToolsGolden(
           '../test_infra/goldens/codeview_scrollbars.png',
         ),
