+++ packages/devtools_app/test/debugger/debugger_screen_breakpoints_test.dart
@@ -74,7 +74,7 @@ void main() {
   ) async {
     await tester.pumpWidget(
       wrapWithControllers(
-        const DebuggerScreenBody(),
+        const DebuggerWindows(),
         debugger: controller,
       ),
     );
