+++ packages/devtools_app/test/debugger/debugger_screen_call_stack_test.dart
@@ -51,7 +51,7 @@ void main() {
   ) async {
     await tester.pumpWidget(
       wrapWithControllers(
-        const DebuggerScreenBody(),
+        const DebuggerWindows(),
         debugger: controller,
       ),
     );
