+++ packages/devtools_app/test/debugger/debugger_screen_dap_variables_test.dart
@@ -61,7 +61,7 @@ void main() {
   ) async {
     await tester.pumpWidget(
       wrapWithControllers(
-        const DebuggerScreenBody(),
+        const DebuggerWindows(),
         debugger: controller,
       ),
     );
